extract create_alias_job function
