+++ backend/es/es_alias.py
@@ -1,52 +1,9 @@
-#!/usr/bin/env python3
-from django.db import transaction
 from typing import Generator
 
-from addcorpus.models import Corpus, CorpusConfiguration
-from ianalyzer.elasticsearch import elasticsearch, server_for_corpus, client_from_config
+from addcorpus.models import CorpusConfiguration
+from ianalyzer.elasticsearch import client_from_config
 from es.models import Server, Index
-from es.sync import update_server_table_from_settings
-from indexing.models import IndexJob, DeleteIndexTask, RemoveAliasTask, AddAliasTask
-from es.versioning import highest_version_in_result, version_from_name
-
-import logging
-logger = logging.getLogger('indexing')
-
-@transaction.atomic
-def create_alias_job(corpus: Corpus, clean=False) -> IndexJob:
-    '''
-    Create a job to move the alias of a corpus to the index with the highest version
-    '''
-
-    job = IndexJob.objects.create(corpus=corpus)
-
-    corpus_config = corpus.configuration
-    corpus_name = corpus.name
-    update_server_table_from_settings()
-    server = Server.objects.get(name=server_for_corpus(corpus_name))
-    index_name = corpus_config.es_index
-    index_alias = corpus_config.es_alias
-    client = elasticsearch(corpus_name)
-
-    alias = index_alias if index_alias else index_name
-    indices = client.indices.get(index='{}-*'.format(index_name))
-    highest_version = highest_version_in_result(indices, alias)
-
-    for index_name, properties in indices.items():
-        is_aliased = alias in properties['aliases'].keys()
-        is_highest_version = version_from_name(index_name, alias) == highest_version
-        index, _ = Index.objects.get_or_create(server=server, name=index_name)
-
-        if not is_highest_version and clean:
-            DeleteIndexTask.objects.create(job=job, index=index)
-
-        if not is_highest_version and is_aliased and not clean:
-            RemoveAliasTask.objects.create(job=job, index=index, alias=alias)
-
-        if is_highest_version and not is_aliased:
-            AddAliasTask.objects.create(job=job, index=index, alias=alias)
-
-    return job
+from indexing.models import DeleteIndexTask, RemoveAliasTask, AddAliasTask
 
 
 def add_alias(task: AddAliasTask):
