+++ backend/indexing/management/commands/alias.py
@@ -1,7 +1,7 @@
 from django.core.management import BaseCommand
 
 from addcorpus.models import Corpus
-from es.es_alias import create_alias_job
+from indexing.create_job import create_alias_job
 from es.es_index import perform_indexing
 
 class Command(BaseCommand):
